+++ src/mympd_api/settings.h
@@ -11,6 +11,7 @@
 #include "../lib/mympd_state.h"
 #include "../lib/validate.h"
 
+bool settings_to_webserver(struct t_mympd_state *mympd_state);
 void mympd_api_settings_statefiles_global_read(struct t_mympd_state *mympd_state);
 void mympd_api_settings_statefiles_partition_read(struct t_partition_state *partition_state);
 sds mympd_api_settings_get(struct t_partition_state *partition_state, sds buffer, long request_id);
@@ -18,5 +19,5 @@ bool mympd_api_settings_cols_save(struct t_mympd_state *mympd_state, sds table, 
 bool mympd_api_settings_set(sds key, sds value, int vtype, validate_callback vcb, void *userdata, sds *error);
 bool mympd_api_settings_mpd_options_set(sds key, sds value, int vtype, validate_callback vcb, void *userdata, sds *error);
 bool mympd_api_settings_connection_save(sds key, sds value, int vtype, validate_callback vcb, void *userdata, sds *error);
-bool mympd_api_settings_partition_save(struct t_partition_state *partition_state, sds color);
+bool mympd_api_settings_partition_set(sds key, sds value, int vtype, validate_callback vcb, void *userdata, sds *error);
 #endif