+++ src/web_server/utility.h
@@ -29,7 +29,7 @@ struct t_mg_user_data {
     bool publish_playlists;      //!< true if mpd playlist directory is configured
     bool publish_music;          //!< true if mpd music directory is accessable
     int connection_count;        //!< number of http connections
-    sds stream_uri;              //!< uri for the mpd stream reverse proxy
+    struct t_list stream_uris;   //!< uri for the mpd stream reverse proxy
     struct t_list session_list;  //!< list of myMPD sessions (pin protection mode)
 };